Preparations for Oomph setup.
Added a new property to the build.properties file of all editor support
and graph layout algorithm plug-ins. The property identifies both kinds
of projects, which will be used by the Oomph setup to setup working
sets. Another possibility would have been to change the plug-in names,
but that would have required package names and everything to be changed
yet again, which is quite a bit of work.