layered.splines: Improved layer spacing in sloppy spline routing
When using sloppy spline routing some additinoal space between layers
can be desired to produce better curves. The new option introduces a
spacing factor to limit the steepness between layers.
The calculation of the space between layers is adapted from the polyline
routing.

In addition a bug in the spline router is fixed, that caused improper
routing when using bignodes.

Signed-off-by: Nis Wechselberg <enbewe@enbewe.de>