Small build restructuring.
A problem with the former state of the build was that it was not well suited for producing Maven repositories because of the parent pom structure. The artifacts we want to publish reference the plugins/pom.xml, which is there for convenience, but was never meant to be published. They now reference the proper main build parent pom. Also, duplicate plugin configurations were removed.