layered.splines: Reduced spline edge/node intersections
With sloppy spline routing there were some cases were edges could
intersect with nodes. This happened when nodes in one layer had
different sizes. The change introduces additional control point at the
start or end of the layer if the node does not cover the complete width
of the layer.

Signed-off-by: Nis Wechselberg <enbewe@enbewe.de>