layered: Fixed unexpected port anchor behaviour. #81
A custom set port anchor was previously not adhered to if the port was placed east or west. This is now fixed by… well, fixing the port anchor if it was set by the user. The port anchor is only adapted to the port side it ends up in if the user has not set an explicit port anchor.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>