core.meta: Implemented documentation generator for *.melk files #91
The MelkDocumentationGenerator creates a Markdown file for each algorithm, group and option in a bundle.
These files contain documentation like descriptions, identifiers, supported options, default values, dependencies etc.
Furthermore it is possible to specify additional documentation for any bundle member via a 'documentation' field.
The generated files reside in the the directory 'docs' of the elk repository. GitHub utilizes the static-site-generator Jekyll
to compile the Markdown files into a website.

Signed-off-by: Daniel Grevismuehl <dag@informatik.uni-kiel.de>