layoutview: Changed layout option collection from lists to set
The layout view created ArrayIndexOutOfBoundsExceptions when changing
the selection in diagrams. This was caused by us creating a list of all
available layout options and therefore creating duplicate entries when
multiple targets were allowed. The PropertyView uses sets for the
properties and was creating the exceptions.
This change gets rid of the lists and uses a set instead to prepare the
layout options for the property view.

Signed-off-by: Nis Wechselberg <enbewe@enbewe.de>