Build: Changed dependencies and cleaned up versions.
The main build file had version numbers distributed throughout the file. They were consolidated into one property.

Also, the meta data compiler had to build the core plug-in. This is not necessary anymore, thereby eliminating a cyclic dependency where the meta data compiler needs the core plug-in, which in turn needs the meta data compiler to be successfully compiled.