Core, Layered: Changed interpretation of "compound node". #219
Graph adapters previously only considered nodes to be compound nodes
that had children. This change also considers nodes as compound nodes if
they hafe inside self loops enabled. We try to get away with this
interpretation, but it is possible that we need to narrow that down to
nodes that have inside self loops enabled _and_ that actually _have_
inside self-loops.