Core, Layered: Inside port labels next to hierarchical ports. #219
By default, inside port labels of hierarchical nodes are always placed
with an offset so as not to be crossed by edges incident to the port.
That offset is not necessary if a port does not have any inside
connections. This commit adds a new layout option to have inside labels
of such ports be placed next the port, without the offset.