Core: Added space-efficient outside port label placement. #221
This commit adds a new size option, SPACE_EFFICIENT_PORT_LABELS. If
activated, the label of the leftmost port of horizontal port sides / the
topmost port of vertical port sides is placed to the port's left / above
the port. This behavior was previously only active if a port side
consisted of only two nodes, but can now be activated also if a port
side has more.