Core, Layered: Calculate minimal size for hierarchical nodes. #182
Hierarchical nodes could end up being too small to place their ports and things. This commit adds a flag to the node size calculation code to not apply the results, but instead return the size the code would like to apply to the node. ELK Layered uses that value as the minimum size to be applied to a hierarchical node. The result is actually too large now, but that will be fixed in a separate commit.