Layered: Fix inside self loops with ports on the same side. #297
This commit relaxes the constraints put on in-layer edges a bit.
Usually, we expected in-layer edges to always connect a regular node to
a dummy node (that's how we handle inverted ports, which was the reason
for introducing in-layer edges in the first place). With this commit,
they can also connect two external port dummies.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>