alg.layered: Implemented predicates for NorthSouth and LongEdge Dummies
As a preliminary step to be able to merge north/south dummies with long
edge dummies (#328), this commit replaces most (not all) checks for the
enum values with the corresponding predicates.
There are still some places left where the enum values are used
directly, primarily in some switch/cases in the crossing minimizer and
in the spacing preparation.

Signed-off-by: Nis Wechselberg <enbewe@enbewe.de>