Docs: Generate additional reference metadata to fix wrong IDs. #354
Formerly, the Hugo website used document identifiers to display the IDs
of algorithms, options, and groups. Since #313 was fixed, the IDs used
for website pages doesn't match the original IDs of the objects anymore.
This commit fixes this problem by adding the original ID as another
property to each page.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>