Core: Major overhaul of debugging infrastructure (#292)
This commits extends the progress monitors such that they can be used to log debug messages and intermediate graphs. Instead of using `System.out.println(...)`, the recommended way to produce debug output is now `monitor.log(...)`. The messages are collected and can be shown in the UI as well as written to debug files. Also, intermediate graphs can be logged with progress monitors and written to debug graphs.

To adapt the UI to the new debugging infrastructure, we add a new view to display debug messages and extend the existing views to show more information and be more user-friendly.