Tests: Improve test descriptions.
Jenkins seems to have a problem with some of our test descriptions. Just
in case that might have something to do with the square braces used in
test descriptions, I'm changing those to parantheses.

Also, this commit fixes the textual description of TestConfiguration,
which failed to mention whether addEdgeLabels was active or not.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>