Test: Register meta data providers explicitly again. #402
There seems to be a problem with our Tycho-Surefire test harness. The
service loader does not see the service provider definitions, as is the
case when we start an Eclipse-based runtime. However, the ELK Service
plug-in probably is not loaded either and thus cannot fix the problem.
Instead of fixing our test harness, we simply revert to registering the
meta data providers explicitly, which is way easier.

Signed-off-by: Christoph Daniel Schulze <cds@informatik.uni-kiel.de>