layered: Order outgoing ports by the node they connect to #608.
If two ports of a node have the same long edge target node. They should
be ordered next to each other because any other ordering would produce
unnecessary crossings. This is done by using the minimal order of the
ports connecting to the same node as the model order for all of them.

Signed-off-by: Soeren Domroes <sdo@informatik.uni-kiel.de>