#701: Add tests concerning port labels with fixed location
This commit adds new tests concerning the labels location for each kind
of org.eclipse.elk.core.options.PortLabelPlacement (INSIDE, OUTSIDE and
"FIXED"). This allows to compare the result for "FIXED" that was not
correctly handled (in comparison to INSIDE and/or OUTSIDE).

These tests also check that the spacing is correctly managed according
to these ports labels.

Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>