Layered: Model order overrules priority of components.
Previously components would be ordered by the summed up priority of each
component and if no priority was set or it was equal, they were ordered
by model order. Since a summed up priority requires the user to
calculate the separate components to set this sensibly it is now
overruled by the model order.

Signed-off-by: Soeren Domroes <sdo@informatik.uni-kiel.de>