Consider node's size when computing available space in port placement (#955)
* Consider node's size when computing available space in port placement

* fix typo

* expand comments

* Update test

Let test also allow node's set height as correct if it is greater than
the the minimum height needed according to the ports. This may not be
the desired solution.

* Add top-down property to nodeContext

Only consider existing node height in the top-down layout case

* Revert changes to test case

* Add test

Tests that in the bottom-up case and in the top-down case
port positions are calculated as expected.